Fixed bug that newly created task is not shown.
Refactored HomePage to use FutureBuilder so it can show if it is loading or if it found data. Fixed problem with not showing newly created task by using navigator.push instead of pop. Needed to add automaticallyImplyLeading: false so that back arrow is not shown in appbar on home screen.