Changed date selection for create and edit task.
Now it is only possible to select today for the task start when a task is created as it should not be possible to plan tasks in the future. The end can be set to today or tomorrow because it should be possible to also work overnight. When editing task the behaviour is the same but for the original dates of the task in the past.