Implemented global state for selected date on home page.
Create TaskFilterProvider to store selectedDate from home page as global state. Using the value from the global state in the home page. Added setting the global state to today if a new task is created because the date on the home page can be different when the floating button is pressed. Added provider to tests.