Added error handling in the view layer.
Added error page. Added try catch for the cases where a common task name or a task is edited or deleted and implemented navigating to the error page in the catch block. Added navigating to the error page when a invalid page index is selected which should probably be impossible to do.